+++ gcn_classic_to_json/notices/SWIFT_POINTDIR/__init__.py
@@ -7,9 +7,9 @@ def parse(bin):
     bin[17:19]  # Spare. According to Docs: '8 bytes for the future'
     bin[20:22]  # Spare. According to Docs: '8 bytes for the future'
 
-    _, bat_mode = bin[11:12].view(dtype=">i2")
-    _, xrt_mode = bin[12:13].view(dtype=">i2")
-    _, uvot_mode = bin[13:14].view(dtype=">i2")
+    _, bat_mode = bin[11:12].view(dtype="<i2")
+    _, xrt_mode = bin[12:13].view(dtype="<i2")
+    _, uvot_mode = bin[13:14].view(dtype="<i2")
 
     return {
         **parse_pointdir(bin),
