+++ gcn_classic_to_json/notices/SWIFT_SC_SLEW/__init__.py
@@ -18,7 +18,7 @@ def parse_slew(bin):
     record_number = np.packbits(np.flip(id_record_number_bits[24:]))
     id = id_record_number_bits[:24].dot(2 ** np.arange(24))
 
-    lat, lon = bin[10:11].view(dtype=">i2")
+    lat, lon = bin[10:11].view(dtype="<i2")
 
     integ_time = bin[15] * 4 / 1000
